+++ src/commands/create.js
@@ -11,6 +11,7 @@ var Promise = require('bluebird'),
 	validatePackage = require('../tasks/validate-package'),
 	retriableWrap = require('../util/wrap'),
 	rebuildWebApi = require('../tasks/rebuild-web-api'),
+	readjson = require('../util/readjson'),
 	fs = Promise.promisifyAll(require('fs'));
 module.exports = function create(options) {
 	'use strict';
@@ -31,9 +32,6 @@ module.exports = function create(options) {
 			});
 		},
 		validationError = function () {
-			if (!options.name) {
-				return 'project name is missing. please specify with --name';
-			}
 			if (!options.region) {
 				return 'AWS region is missing. please specify with --region';
 			}
@@ -59,10 +57,22 @@ module.exports = function create(options) {
 				return 'no files match additional policies (' + options.policies + ')';
 			}
 		},
-		createLambda = function (zipFile, roleArn, retriesLeft) {
+		getPackageName = function () {
+			if (options.name) {
+				return Promise.resolve(options.name);
+			}
+			return readjson(path.join(source, 'package.json')).then(function (jsonConfig) {
+				var name = jsonConfig.name && jsonConfig.name.trim();
+				if (!name) {
+					return Promise.reject('project name is missing. please specify with --name or in package.json');
+				}
+				return name;
+			});
+		},
+		createLambda = function (functionName, zipFile, roleArn, retriesLeft) {
 			var functionMeta = {
 				Code: { ZipFile: zipFile },
-				FunctionName: options.name,
+				FunctionName: functionName,
 				Handler: options.handler || (options['api-module'] + '.router'),
 				Role: roleArn,
 				Runtime: options.runtime || 'nodejs4.3',
@@ -76,7 +86,7 @@ module.exports = function create(options) {
 			return lambda.createFunctionPromise(functionMeta).catch(function (error) {
 				if (error && error.cause && error.cause.message == iamPropagationError) {
 					return Promise.delay(3000).then(function () {
-						return createLambda(zipFile, roleArn, retriesLeft - 1);
+						return createLambda(functionName, zipFile, roleArn, retriesLeft - 1);
 					});
 				} else {
 					return Promise.reject(error);
@@ -101,7 +111,7 @@ module.exports = function create(options) {
 				return Promise.reject('No apiConfig defined on module \'' + options['api-module'] + '\'. Are you missing a module.exports?');
 			}
 			return apiGateway.createRestApiAsync({
-				name: options.name
+				name: lambdaMetadata.FunctionName
 			}).then(function (result) {
 				lambdaMetadata.api = {id: result.id, module: options['api-module']};
 				return rebuildWebApi(lambdaMetadata.FunctionName, options.version || 'latest', result.id, apiConfig, options.region, options.verbose);
@@ -128,14 +138,14 @@ module.exports = function create(options) {
 				return config;
 			});
 		},
-		loadRole = function () {
+		loadRole = function (functionName) {
 			if (options.role) {
 				return iam.getRoleAsync({RoleName: options.role});
 			} else {
 				return fs.readFileAsync(templateFile('lambda-exector-policy.json'), 'utf8')
 					.then(function (lambdaRolePolicy) {
 						return iam.createRoleAsync({
-							RoleName: options.name + '-executor',
+							RoleName: functionName + '-executor',
 							AssumeRolePolicyDocument: lambdaRolePolicy
 						});
 					});
@@ -147,18 +157,22 @@ module.exports = function create(options) {
 				return addPolicy(policyName, roleMetadata.Role.RoleName, fileName);
 			});
 		},
-		packageArchive;
+		packageArchive,
+		functionName;
 	if (validationError()) {
 		return Promise.reject(validationError());
 	}
-	return collectFiles(source)
-	.then(function (dir) {
+	return getPackageName().then(function (name) {
+		functionName = name;
+	}).then(function () {
+		return collectFiles(source);
+	}).then(function (dir) {
 		return validatePackage(dir, options.handler, options['api-module']);
-	})
-	.then(zipdir).then(function (zipFile) {
+	}).then(zipdir).then(function (zipFile) {
 		packageArchive = zipFile;
-	}).then(loadRole)
-	.then(function (result) {
+	}).then(function () {
+		return loadRole(functionName);
+	}).then(function (result) {
 		roleMetadata = result;
 	}).then(function () {
 		return addPolicy('log-writer', roleMetadata.Role.RoleName);
@@ -169,7 +183,7 @@ module.exports = function create(options) {
 	}).then(function () {
 		return fs.readFileAsync(packageArchive);
 	}).then(function (fileContents) {
-		return createLambda(fileContents, roleMetadata.Role.Arn, 10);
+		return createLambda(functionName, fileContents, roleMetadata.Role.Arn, 10);
 	})
 	.then(function (lambdaMetadata) {
 		if (options['api-module']) {
@@ -186,20 +200,9 @@ module.exports.doc = {
 	priority: 1,
 	args: [
 		{
-			argument: 'name',
-			description: 'lambda function name',
-			example: 'awesome-microservice'
-		},
-		{
 			argument: 'region',
 			description: 'AWS region where to create the lambda',
 			example: 'us-east-1'
-		},
-		{
-			argument: 'version',
-			optional: true,
-			description: 'A version alias to automatically assign to the new function',
-			example: 'development'
 		},
 		{
 			argument: 'handler',
@@ -214,6 +217,19 @@ module.exports.doc = {
 				'If you provide this parameter, the handler option is ignored.\n' +
 				'This should be a module created using the Claudia API Builder.',
 			example: 'if the api is defined in web.js, this would be web'
+		},
+		{
+			argument: 'name',
+			optional: true,
+			description: 'lambda function name',
+			example: 'awesome-microservice',
+			'default': 'the project name from package.json'
+		},
+		{
+			argument: 'version',
+			optional: true,
+			description: 'A version alias to automatically assign to the new function',
+			example: 'development'
 		},
 		{
 			argument: 'source',